requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
Flask==2.2.5
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
greenlet==3.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
openai==0.28.1
protobuf==3.20.3
python-decouple==3.8
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.22
telegram==0.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.6
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
